PR feedback; restructure
